+++ src/UIDMotionBlur.cpp
@@ -42,7 +42,7 @@ void UIDMotionBlur::on_resize(int width, int height) {
 
 void UIDMotionBlur::process_event(OUI* element, uint32_t message, uint64_t param, bool bubbleUp) {
 	if (element == &cSize) {
-		size = cSize.get_value();
+		size = (size_t)cSize.get_value();
 		shouldInvalidate = true;
 	}
 	else if (element == &slAngle) {
